tests/test_simulator.py
```python
import unittest
from simulator import Simulator
from multiverse import Multiverse
from universe import Universe

class TestSimulator(unittest.TestCase):
    def setUp(self):
        """
        Set up a new Simulator instance before each test.
        """
        self.simulator = Simulator()

    def test_add_universe(self):
        """
        Test that a universe can be added to the simulator.
        """
        properties = {"property1": "value1", "property2": "value2", "property3": "value3"}
        self.simulator.add_universe("universe1", properties)
        self.assertIsInstance(self.simulator.multiverse.get_universe("universe1"), Universe)

    def test_remove_universe(self):
        """
        Test that a universe can be removed from the simulator.
        """
        properties = {"property1": "value1", "property2": "value2", "property3": "value3"}
        self.simulator.add_universe("universe1", properties)
        self.simulator.remove_universe("universe1")
        self.assertIsNone(self.simulator.multiverse.get_universe("universe1"))

    def test_get_universe_properties(self):
        """
        Test that the properties of a universe can be retrieved from the simulator.
        """
        properties = {"property1": "value1", "property2": "value2", "property3": "value3"}
        self.simulator.add_universe("universe1", properties)
        self.assertEqual(self.simulator.get_universe_properties("universe1"), properties)

    def test_set_universe_properties(self):
        """
        Test that the properties of a universe can be set in the simulator.
        """
        properties = {"property1": "value1", "property2": "value2", "property3": "value3"}
        new_properties = {"property1": "new_value1", "property2": "new_value2", "property3": "new_value3"}
        self.simulator.add_universe("universe1", properties)
        self.simulator.set_universe_properties("universe1", new_properties)
        self.assertEqual(self.simulator.get_universe_properties("universe1"), new_properties)

    def test_simulate(self):
        """
        Test that the simulate method works correctly. This test will need to be updated
        once the simulate method is implemented.
        """
        properties = {"property1": "value1", "property2": "value2", "property3": "value3"}
        self.simulator.add_universe("universe1", properties)
        self.simulator.simulate()
        # TODO: Add assertions to check the result of the simulation

if __name__ == "__main__":
    unittest.main()
```
